pyproject.toml
[tool.poetry]
name = "argus"
version = "0.1.0"
description = "Campaign Intelligence System — GraphRAG backend"
authors = []
packages = [
    { include = "argus" },
    { include = "scripts" },
]

[tool.poetry.dependencies]
python = "^3.9"

# Web
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.30"}
sse-starlette = "^2.1"
python-multipart = "^0.0.9"

# Pipeline
langgraph = "^0.2"
langchain-openai = "^0.2"
langchain-postgres = "^0.0.12"
langchain-neo4j = "^0.1"

# Storage
psycopg = {extras = ["binary"], version = "^3.1"}
asyncpg = "^0.29"
pgvector = ">=0.2.5,<0.3.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
neo4j = "^5.24"
redis = {extras = ["hiredis"], version = "^5.0"}
arq = "^0.26"

# ML / Embeddings
sentence-transformers = "^3.0"
einops = "^0.8"
spacy = "^3.7"
pypdf = "^4.0"
tiktoken = "^0.7"

# Config
pydantic-settings = "^2.0"
pyyaml = "^6.0"
deepmerge = "^2.0"
injector = "^0.21"

# Auth
pyjwt = {extras = ["cryptography"], version = "^2.9"}

# Observability
langfuse = "^2.0"

# Rate limiting
slowapi = "^0.1"

# Seed script deps
click = "^8.1"
pandas = "^2.2"
faker = "^24.0"
openpyxl = "^3.1"
reportlab = "^4.1"
python-docx = "^1.1"
jinja2 = "^3.1"
kaggle = "^1.6"
python-dotenv = "^1.0"
rich = "^13.7"
matplotlib = "^3.9"

[tool.poetry.scripts]
argus      = "argus.__main__:main"
seed       = "scripts.seed_data:cli"
download-kaggle = "scripts.download_kaggle:cli"

# Week 10 — add when building guardrails component:
# llm-guard = {version = "^0.3", python = ">=3.9,<3.12"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
httpx = "^0.27"
ruff = "^0.6"
mypy = "^1.11"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
